api/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.3
cffi==2.0.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.7
fastapi==0.135.3
gevent==25.9.1
greenlet==3.3.2
gunicorn==25.3.0
h11==0.16.0
idna==3.11
inflect==7.5.0
iniconfig==2.3.0
kombu==5.6.2
more-itertools==11.0.2
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
sqlacodegen==4.0.3
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
typeguard==4.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uvicorn==0.43.0
vine==5.1.0
wcwidth==0.6.0
zope.event==6.1
zope.interface==8.2
